fastapi-solution/etl/requirements.txt
python-dotenv==0.21.1
psycopg2-binary==2.9.5
elasticsearch~=7.15.1
pydantic~=1.8.2